+++ core-ng/src/test/java/core/framework/impl/cache/CacheImplTest.java
@@ -1,6 +1,7 @@
 package core.framework.impl.cache;
 
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
@@ -13,12 +14,18 @@ import static org.mockito.Mockito.when;
  * @author neo
  */
 public class CacheImplTest {
+    CacheStore cacheStore;
+
+    @Before
+    public void prepare() {
+        cacheStore = Mockito.mock(CacheStore.class);
+    }
+
     @Test
     public void get() {
-        CacheStore store = Mockito.mock(CacheStore.class);
-        when(store.get("name", "key")).thenReturn("1");
+        when(cacheStore.get("name", "key")).thenReturn("1");
 
-        CacheImpl<Integer> cache = new CacheImpl<>("name", Integer.class, Duration.ofHours(1), store);
+        CacheImpl<Integer> cache = new CacheImpl<>("name", Integer.class, Duration.ofHours(1), cacheStore);
 
         Integer value = cache.get("key", () -> null);
         Assert.assertEquals(1, (int) value);
@@ -26,12 +33,9 @@ public class CacheImplTest {
 
     @Test
     public void put() {
-        CacheStore store = Mockito.mock(CacheStore.class);
-        when(store.get("name", "key")).thenReturn("1");
-
-        CacheImpl<Integer> cache = new CacheImpl<>("name", Integer.class, Duration.ofHours(1), store);
+        CacheImpl<Integer> cache = new CacheImpl<>("name", Integer.class, Duration.ofHours(1), cacheStore);
         cache.put("key", 1);
 
-        verify(store).put("name", "key", "1", Duration.ofHours(1));
+        verify(cacheStore).put("name", "key", "1", Duration.ofHours(1));
     }
 }