+++ core-ng/src/test/java/core/framework/impl/web/service/ServiceControllerBuilderTest.java
@@ -7,13 +7,15 @@ import core.framework.api.web.Controller;
 import core.framework.api.web.Request;
 import core.framework.api.web.Response;
 import core.framework.api.web.ResponseImpl;
-import core.framework.api.web.service.PathParam;
 import core.framework.impl.web.response.BeanBody;
-import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
 import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.when;
 
 /**
  * @author neo
@@ -22,7 +24,7 @@ public class ServiceControllerBuilderTest {
     public static class TestWebServiceImpl implements TestWebService {
         @Override
         public TestResponse get(Integer id) {
-            Assert.assertEquals(1, (int) id);
+            assertEquals(1, (int) id);
 
             TestResponse response = new TestResponse();
             response.intField = 2;
@@ -30,37 +32,44 @@ public class ServiceControllerBuilderTest {
         }
 
         @Override
-        public void create(@PathParam("id") Integer id, TestRequest request) {
-            Assert.assertEquals(1, (int) id);
-            Assert.assertEquals("value", request.stringField);
+        public void create(Integer id, TestRequest request) {
+            assertEquals(1, (int) id);
+            assertEquals("value", request.stringField);
         }
 
         @Override
-        public void delete(@PathParam("id") String id) {
+        public void delete(String id) {
 
         }
 
         @Override
         public List<TestResponse> batch(List<TestRequest> requests) {
-            Assert.assertEquals(1, requests.size());
+            assertEquals(1, requests.size());
             return Lists.newArrayList();
         }
     }
 
-    TestWebServiceImpl serviceImpl = new TestWebServiceImpl();
+    Request request;
+    TestWebServiceImpl serviceImpl;
+
+    @Before
+    public void prepare() {
+        serviceImpl = new TestWebServiceImpl();
+        request = Mockito.mock(Request.class);
+    }
 
     @Test
     public void get() throws Exception {
-        Request request = Mockito.mock(Request.class);
-        Mockito.when(request.pathParam("id", Integer.class)).thenReturn(1);
+        when(request.pathParam("id", Integer.class)).thenReturn(1);
 
         Controller controller = new ServiceControllerBuilder<>(TestWebService.class,
             serviceImpl,
             TestWebService.class.getDeclaredMethod("get", Integer.class))
             .build();
+
         Response response = controller.execute(request);
-        Assert.assertEquals(HTTPStatus.OK, response.status());
-        Assert.assertEquals(2, (int) ((TestWebService.TestResponse) ((BeanBody) ((ResponseImpl) response).body).bean).intField);
+        assertEquals(HTTPStatus.OK, response.status());
+        assertEquals(2, (int) ((TestWebService.TestResponse) ((BeanBody) ((ResponseImpl) response).body).bean).intField);
     }
 
     @Test
@@ -68,16 +77,15 @@ public class ServiceControllerBuilderTest {
         TestWebService.TestRequest requestBean = new TestWebService.TestRequest();
         requestBean.stringField = "value";
 
-        Request request = Mockito.mock(Request.class);
-        Mockito.when(request.pathParam("id", Integer.class)).thenReturn(1);
-        Mockito.when(request.bean(TestWebService.TestRequest.class)).thenReturn(requestBean);
+        when(request.pathParam("id", Integer.class)).thenReturn(1);
+        when(request.bean(TestWebService.TestRequest.class)).thenReturn(requestBean);
 
         Controller controller = new ServiceControllerBuilder<>(TestWebService.class,
             serviceImpl,
             TestWebService.class.getDeclaredMethod("create", Integer.class, TestWebService.TestRequest.class))
             .build();
         Response response = controller.execute(request);
-        Assert.assertEquals(HTTPStatus.CREATED, response.status());
+        assertEquals(HTTPStatus.CREATED, response.status());
     }
 
     @Test
@@ -85,14 +93,13 @@ public class ServiceControllerBuilderTest {
         TestWebService.TestRequest requestBean = new TestWebService.TestRequest();
         requestBean.stringField = "value";
 
-        Request request = Mockito.mock(Request.class);
-        Mockito.when(request.bean(Types.list(TestWebService.TestRequest.class))).thenReturn(Lists.newArrayList(requestBean));
+        when(request.bean(Types.list(TestWebService.TestRequest.class))).thenReturn(Lists.newArrayList(requestBean));
 
         Controller controller = new ServiceControllerBuilder<>(TestWebService.class,
             serviceImpl,
             TestWebService.class.getDeclaredMethod("batch", List.class))
             .build();
         Response response = controller.execute(request);
-        Assert.assertEquals(HTTPStatus.OK, response.status());
+        assertEquals(HTTPStatus.OK, response.status());
     }
 }