+++ core-ng/src/test/java/core/framework/impl/web/service/WebServiceClientBuilderTest.java
@@ -2,46 +2,59 @@ package core.framework.impl.web.service;
 
 import core.framework.api.http.HTTPMethod;
 import core.framework.api.util.Maps;
-import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import java.lang.reflect.Type;
 import java.util.Map;
 
-import static org.mockito.Matchers.anyString;
+import static org.junit.Assert.assertSame;
 import static org.mockito.Matchers.eq;
-import static org.mockito.Matchers.isNull;
+import static org.mockito.Matchers.startsWith;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
  * @author neo
  */
 public class WebServiceClientBuilderTest {
+    WebServiceClient webServiceClient;
+
+    @Before
+    public void prepare() {
+        webServiceClient = Mockito.mock(WebServiceClient.class);
+    }
+
     @Test
     public void get() {
-        WebServiceClient webServiceClient = Mockito.mock(WebServiceClient.class);
         Map<String, String> pathParams = Maps.newHashMap();
         pathParams.put("id", "1");
-        when(webServiceClient.execute(eq(HTTPMethod.GET), anyString(), eq(pathParams), isNull(Type.class), isNull(), eq(TestWebService.TestResponse.class)))
-            .thenReturn(new TestWebService.TestResponse());
+        TestWebService.TestResponse expectedResponse = new TestWebService.TestResponse();
+
+        when(webServiceClient.serviceURL(startsWith("/test/:id"), eq(pathParams)))
+            .thenReturn("http://localhost/test/1");
+        when(webServiceClient.execute(HTTPMethod.GET, "http://localhost/test/1", null, null, TestWebService.TestResponse.class))
+            .thenReturn(expectedResponse);
 
         TestWebService client = new WebServiceClientBuilder<>(TestWebService.class, webServiceClient).build();
 
         TestWebService.TestResponse response = client.get(1);
-        Assert.assertNotNull(response);
+        assertSame(expectedResponse, response);
     }
 
     @Test
     public void create() {
-        WebServiceClient webServiceClient = Mockito.mock(WebServiceClient.class);
         Map<String, String> pathParams = Maps.newHashMap();
         pathParams.put("id", "1");
-        when(webServiceClient.execute(eq(HTTPMethod.POST), anyString(), eq(pathParams), eq(TestWebService.TestRequest.class), isNull(), eq(Void.class)))
-            .thenReturn(null);
+
+        when(webServiceClient.serviceURL(startsWith("/test/:id"), eq(pathParams)))
+            .thenReturn("http://localhost/test/1");
 
         TestWebService client = new WebServiceClientBuilder<>(TestWebService.class, webServiceClient).build();
 
-        client.create(1, new TestWebService.TestRequest());
+        TestWebService.TestRequest request = new TestWebService.TestRequest();
+        client.create(1, request);
+
+        verify(webServiceClient).execute(HTTPMethod.PUT, "http://localhost/test/1", TestWebService.TestRequest.class, request, void.class);
     }
 }